feat: remove now replaced download functions and factor out edwards magnitude conversion
