+++ catalog_tools/__init__.py
@@ -2,29 +2,17 @@
 # flake8: noqa
 
 # analysis
-from catalog_tools.analysis.estimate_beta import\
-    estimate_b_elst,\
-    estimate_b_utsu,\
-    estimate_beta_tinti,\
-    estimate_b_tinti,\
-    estimate_b_laplace,\
-    shi_bolt_confidence
-
-# download
-from catalog_tools.download.download_catalogs import\
-    download_catalog_sed,\
-    download_catalog_scedc,\
-    download_catalog
-
+from catalog_tools.analysis.estimate_beta import (estimate_b_elst,
+                                                  estimate_b_laplace,
+                                                  estimate_b_tinti,
+                                                  estimate_b_utsu,
+                                                  estimate_beta_tinti,
+                                                  shi_bolt_confidence)
 # plots
-from catalog_tools.plots.basics import\
-    plot_cum_fmd,\
-    plot_fmd,\
-    plot_cum_count,\
-    plot_mags_in_time
+from catalog_tools.plots.basics import (plot_cum_count, plot_cum_fmd, plot_fmd,
+                                        plot_mags_in_time)
 from catalog_tools.plots.seismicity import plot_in_space
-
 # utils
 from catalog_tools.utils.binning import bin_to_precision
-from catalog_tools.utils.simulate_distributions import simulate_magnitudes
 from catalog_tools.utils.filtering import cat_intersect_polygon
+from catalog_tools.utils.simulate_distributions import simulate_magnitudes
