+++ setup.cfg
@@ -27,4 +27,5 @@ dev =
     isort
     pytest
     pytest-cov
+    responses
     tox
